refactor: add LottoSeller
