+++ src/main/java/lotto/controller/LottoController.java
@@ -10,7 +10,7 @@ import java.util.stream.Collectors;
 public class LottoController {
 
     public static void main(String[] args) {
-        Lottos tickets = Lottos.issue(InputView.inputMoney());
+        Lottos tickets = LottoSeller.issueLotto(InputView.inputMoney());
 
         ResultView.printCheckPurchaseMessage(tickets.count());
         ResultView.printLottoTickets(tickets);
